Massive changes throughout; first revision tested on robot.
Many changes were made throughout the program.
This version was tested on the actual 2022 robot, and the drive system,
the cargo intake, the cargo conveyor, and the shooter were all verified to
work, though they may need to be tuned a bit (acceleration, speed, PID
coefficients, etc.).
The robot still does not have the infrared cargo sensors, the USB
videocamera, the limelight camera, or either of the climber systems
installed, so none of those have been tested yet.