Many tuning changes, and a couple of fixes.
Changed the 5-ball autonomous maneuver sequence to use the limelight
for aiming and distance calculation when shooting.
Fixed the previous button numbers for the BUTTON_xxxCLIMBERxxx_PREV #defines.
Fixed a calculation for dDesiredYaw in two places (to divide by PI instead
of multiply by PI).
Changed the aggressivity of turns in DriveToCargo()  (This is a temporary
change; we didn't have time to test it well yet -- we still have to test
and find the best value for how aggressive to make the turns when
tracking cargo).
Cleaned up, fixed, and eliminated some comments, and reformatted some code.
Added code to ensure the shooter motors are turned off when not needed, and
initialize them to make sure they are stopped at the start of every match.
Added BurnFlash() calls for all SparkMax motor controllers, to protect
against bad things happening if we have a brownout and those drive motor
controllers lose their settings (one side would no longer be inverted).
Added comment notes at the bottom of the code, to keep track of the
changes we are making and remind us what testing still needs to be done.